--- src/zks_provider/types/mod.rs
@@ -1,2 +1,0 @@
-pub mod fee;
-pub use fee::Fee;
